requirements.txt
requests>=2.31.0
pandas>=2.1.0
matplotlib>=3.8.0
python-dotenv>=1.0.0
tabulate>=0.9.0
folium>=0.15.0
